+++ contrib/owaspdependencycheck/src/mill/contrib/owaspdependencycheck/OwaspDependencyCheckModule.scala
@@ -26,7 +26,7 @@ trait OwaspDependencyCheckModule extends Module, OfflineSupportModule {
     Seq("--nvdDatafeed", "https://dependency-check.github.io/DependencyCheck_Builder/nvd_cache/")
 
   /**
-   * The worker is version specific, so is another version is required, it can be customized here.
+   * The worker is version specific. If another version is required it can be customized here.
    */
   val owaspDependencyCheckWorker: ModuleRef[OwaspDependencyCheckWorker] =
     ModuleRef(OwaspDependencyCheckWorker)
@@ -73,62 +73,4 @@ trait OwaspDependencyCheckModule extends Module, OfflineSupportModule {
         DependencyCheckResult(Seq.empty, 0)
       }
     }
-}
-
-/**
- * Java Dependency Check, that adds the resolvedRunMvnDeps path to be scanned in the dependency check.
- */
-trait OwaspDependencyCheckJavaModule extends JavaModule, OwaspDependencyCheckModule {
-  override def owaspDependencyCheckFiles: T[Seq[PathRef]] = Task {
-    super.resolvedRunMvnDeps()
-  }
-}
-
-trait OwaspDependencyCheckWorker extends CoursierModule, OfflineSupportModule {
-
-  def dependencyCheckVersion: T[String]
-
-  def dependencyCheckClasspath: T[Seq[PathRef]] = Task {
-    defaultResolver().classpath(
-      Seq(mvn"org.owasp:dependency-check-cli:${dependencyCheckVersion()}")
-    )
-  }
-
-  def worker: Worker[DependencyCheckInstance] = Task.Worker {
-    new DependencyCheckInstance(dependencyCheckClasspath().map(_.path))
-  }
-
-  private[owaspdependencycheck] class DependencyCheckInstance(
-      dependencyCheckClasspath: Seq[os.Path]
-  ) extends AutoCloseable {
-    val classLoader = mill.util.Jvm.createClassLoader(dependencyCheckClasspath)
-    val depencencyCheckCli = classLoader.loadClass("org.owasp.dependencycheck.App")
-    val appConstructor = depencencyCheckCli.getConstructor()
-    val mainMethod = depencencyCheckCli.getMethod("run", classOf[Array[String]])
-
-    def runScan(args: Seq[String]): Int = {
-      val ctxLoader = Thread.currentThread().getContextClassLoader
-      try {
-        Thread.currentThread().setContextClassLoader(classLoader)
-        val app = appConstructor.newInstance()
-        mainMethod.invoke(app, args.to(Array)).asInstanceOf[Int]
-      } finally {
-        Thread.currentThread().setContextClassLoader(ctxLoader)
-      }
-    }
-
-    def close() = {
-      classLoader.close()
-    }
-  }
-
-  override def prepareOffline(all: Flag): Command[Seq[PathRef]] = Task.Command {
-    (super.prepareOffline(all)() ++ dependencyCheckClasspath()).distinct
-  }
-}
-
-object OwaspDependencyCheckWorker extends ExternalModule, OwaspDependencyCheckWorker {
-  def dependencyCheckVersion: T[String] = "12.2.0" // FIXME: source from build
-
-  override lazy val millDiscover = Discover[this.type]
 }
